curl_easy_getinfo(CURLINFO_RESPONSE_CODE, ...) expexts long argument,  fix variable types accordingly

+++ src/httpclient.h
@@ -32,7 +32,7 @@ class HttpClient : public HttpInterface {
   virtual HttpResponse download(const std::string &url, curl_write_callback callback, void *userp);
   virtual void setCerts(const std::string &ca, CryptoSource ca_source, const std::string &cert,
                         CryptoSource cert_source, const std::string &pkey, CryptoSource pkey_source);
-  unsigned int http_code;
+  long http_code;
 
  private:
   /**
